WIP: wrong image order.
The problem that we cannot enque readImage without intrusive logic.

Queue has a guaranteed ordering in terms that the first queued task will
be popped first. But it is not guaranteed that the first queued task
will be executed first. I.e.:
It can happen
 - enqueue: read-frame @0
 - enqueue: read-frame @1
 - enqueue: read-frame @2

 - execute @2 read-frame-0
 - execute @0 read-frame-1
 - execute @1 read-frame-2

 - pop read-frame-1
 - pop read-frame-2
 - pop read-frame-0

We could avoid like having an ordering for execution. So far there is a
limitation with future_spawn - One cannot store the future due to its
type is dependent on the sender type. And type ereasure is not possible.